update last fix on otool, prepare for norme

+++ libft/ft_utils/ft_bzero.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_bzero.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: qdequele <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: quentindequelen <quentindequelen@studen    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/24 18:40:02 by qdequele          #+#    #+#             */
-/*   Updated: 2015/11/27 18:20:45 by qdequele         ###   ########.fr       */
+/*   Updated: 2018/02/23 11:04:53 by quentindequ      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../libft.h"
 
-void	ft_bzero(void *s, size_t n)
+void	ft_bzero(void *s, unsigned int n)
 {
-	size_t i;
+	unsigned int i;
 
 	i = 0;
 	while (i < n)
